refactor(udev): split render.rs into planes/backdrop/debug modules
render.rs mixed plane allocation, node wiring, backdrop compositing,
element ordering and debug tooling in two ~650-line functions that were
both named render_surface. Plane swapchain setup moves to planes.rs
(alloc macros folded into one function), the cross-plane blur composite
to backdrop.rs, and dump/realization/toggle tooling to debug.rs; the
free function is renamed render_output_frame. udev also stops reaching
into workspaces internals: overlay-UI activity and current-workspace
window checks are now Workspaces accessors (removing a panic-capable
index). No behavior change.
